Added the 'Flow' button in the chained_request. Also, fixed the /maintenance/reset/scrape_prep1.py to split the 'generators' and 'process_string' to the ';' char when there are multiple. I updated start_prep2.sh to check that the user is root (bugfix for SELINUX socket policy) and also to be called from everywhere.
git-svn-id: svn+ssh://svn.cern.ch/reps/CMSUserCode/trunk/prep2/nikolas/prep2@425 4525493e-7705-40b1-a816-d608a930855b

+++ maintenance/install_prep2.sh
@@ -150,5 +150,13 @@ cd /home/
 sudo chown -R $user:$group prep2/
 
 # reboot
-echo 'Installation Complete! Rebooting...'
-sudo reboot
+echo 'Installation complete.'
+
+echo "You have to reboot to complete the process. Reboot now? [y/n]:"
+read choice
+
+if [ "$choice" == "y" ]; then
+	sudo reboot
+else
+	echo "Aborting..."
+fi
